generify log hooks, support limited form of string substitution, ensure JSON hooks are installed before store setup
